uint -> uint256

+++ src/Goerli-DssSpell.t.base.sol
@@ -1055,10 +1055,10 @@ contract GoerliDssSpellTestBase is Config, DSTest, DSMath {
     function getSignatures(bytes32 signHash) internal returns (bytes memory signatures, address[] memory signers) {
         // seeds chosen s.t. corresponding addresses are in ascending order
         uint8[30] memory seeds = [8,10,6,2,9,15,14,20,7,29,24,13,12,25,16,26,21,22,0,18,17,27,3,28,23,19,4,5,1,11];
-        uint numSigners = seeds.length;
+        uint256 numSigners = seeds.length;
         signers = new address[](numSigners);
-        for(uint i; i < numSigners; i++) {
-            uint sk = uint(keccak256(abi.encode(seeds[i])));
+        for(uint256 i; i < numSigners; i++) {
+            uint256 sk = uint256(keccak256(abi.encode(seeds[i])));
             signers[i] = hevm.addr(sk);
             (uint8 v, bytes32 r, bytes32 s) = hevm.sign(sk, signHash);
             signatures = abi.encodePacked(signatures, r, s, v);
